infrastructure/utils/backend/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
google-cloud-storage==2.13.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click>=8.0.0
MarkupSafe==2.1.3
requests==2.31.0
PyYAML==6.0.1
python-dotenv==1.0.0
prometheus-client==0.17.1
